add scrape/parse jobs for scraper; Dockerfile
